Fetch ClickSmart gripper node name from the environment
We're going to modify our environment setup scripts to grab the
node name of the ClickSmart gripper out of the robot_description URDF
(`rosparam get /robot_description`) and stick it in an environment
variable (SAWYER_GRIPPER_NODENAME).  Modify the Gripper class to take
the node name from this environment variable instead of hard-coding it.